refactor(shared): type background apply event
