+++ plugins/plugin-app-control/src/actions/background.test.ts
@@ -1,6 +1,8 @@
 import type { IAgentRuntime, Media, Memory } from "@elizaos/core";
+import { BACKGROUND_APPLY_EVENT as SHARED_BACKGROUND_APPLY_EVENT } from "@elizaos/shared/events";
 import { describe, expect, it, vi } from "vitest";
 import {
+	BACKGROUND_APPLY_EVENT,
 	type BackgroundApplyPayload,
 	createBackgroundAction,
 	inferBackgroundPlan,
@@ -292,6 +294,10 @@ describe("programmable GLSL shader plan (#10694)", () => {
 });
 
 describe("BACKGROUND action handler", () => {
+	it("uses the shared background apply event contract", () => {
+		expect(BACKGROUND_APPLY_EVENT).toBe(SHARED_BACKGROUND_APPLY_EVENT);
+	});
+
 	function setup() {
 		const emitted: BackgroundApplyPayload[] = [];
 		const replies: string[] = [];
